refactor: extract code audit execution plan
